+++ litesite/templates/default/src/public/main.js
@@ -1,0 +1,9 @@
+;(function () {
+	'use strict'
+
+	const main = () => {
+		document.documentElement.classList.add('js')
+	}
+
+	main()
+})()